Improve start-up speed for new TMUX panes and bootstrap new environments
TMUX uses the global environment as the basis for the environments in
its panels. With this in mind it is possible to compute part of the
environment variables in the global environment and avoid recomputing
those for each panel.

Since the heavy environment checks happen only for the global
environment, this commit includes checks for important components.

Homebrew, Pyenv, and Oh-my-zsh! installation is checked and if these are
missing the user is prompted to install.
Further, some Homebrew packages are checked and installed if missing.

The logic to remove duplicate entries from $PATH was removed in favor of
using the builtin 'typeset' function.

The complex reload logic was reverted.